% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mouse-trial-duplicates.R
\docType{data}
\name{mouse_trial_duplicates}
\alias{mouse_trial_duplicates}
\title{Mouse trial with duplicate records}
\format{
A data frame with 51 rows and the same 7 variables as
  \code{\link{mouse_trial}}.
}
\source{
Derived from \code{\link{mouse_trial}}.
}
\usage{
mouse_trial_duplicates
}
\description{
A shuffled variant of \code{\link{mouse_trial}} with three records repeated.
}
\examples{
data(mouse_trial_duplicates)
mouse_trial_duplicates
}
\keyword{datasets}
