% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-resources.R
\name{course_resources}
\alias{course_resources}
\title{List course resources}
\usage{
course_resources()
}
\value{
A tibble with one row per resource.
}
\description{
Lists the datasets and templates included in the package.
}
