% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-course-project.R
\name{create_course_project}
\alias{create_course_project}
\title{Create a course project}
\usage{
create_course_project(path)
}
\arguments{
\item{path}{Where to create the project.}
}
\value{
`path`, invisibly.
}
\description{
Creates an RStudio and Quarto project for BIOSCI 504 exercises.
}
